lift function to utils
