+++ langchain_hana/utils.py
@@ -23,3 +23,19 @@ def _validate_k_and_fetch_k(k: int, fetch_k: int):
         raise ValueError(
             "Parameter 'fetch_k' must be an integer greater than or equal to 'k'"
         )
+
+
+def _generate_cross_encode_sql_and_params(
+        text_column: str, metadata_column: str, query: str, rank_fields: list[str], rerank_model_id: str
+    ) -> tuple[str, list]:
+        if rank_fields:
+            cross_encode_input = f"'{text_column}:' || TO_NVARCHAR({text_column})"
+            for field in rank_fields:
+                cross_encode_input += f"|| '| {field}:' || TO_NVARCHAR(COALESCE(JSON_VALUE({metadata_column}, '$.{field}'), ''))"
+        else:
+            cross_encode_input = f"TO_NVARCHAR({text_column})"
+
+        cross_encode_sql = f"CROSS_ENCODE({cross_encode_input}, ?, ?) OVER()"
+
+        cross_encode_params = [query, rerank_model_id]
+        return cross_encode_sql, cross_encode_params
